Fix multiplication warnings

+++ engine/src/cmd/armed.cpp
@@ -251,7 +251,7 @@ void Armed::Fire(unsigned int weapon_type_bitmask, bool listen_to_owner) {
                 //&& ( (ROLES::EVERYTHING_ELSE&weapon_type_bitmask&i->type->role_bits) || i->type->role_bits == 0 )
                 ((locked_on && missile_and_want_to_fire_missiles) || gun_and_want_to_fire_guns);
         if ((*i).type->type == WEAPON_TYPE::BEAM) {
-            if ((*i).type->energy_rate * simulation_atom_var > unit->energy.Level()) {
+            if ((*i).type->energy_rate * static_cast<double>(simulation_atom_var) > unit->energy.Level()) {
                 //NOT ONLY IN non-networking mode : anyway, the server will tell everyone including us to stop if not already done
                 (*i).UnFire();
                 continue;
@@ -278,7 +278,7 @@ void Armed::Fire(unsigned int weapon_type_bitmask, bool listen_to_owner) {
                 if (i->type->type == WEAPON_TYPE::BEAM) {
                     if (i->ref.gun) {
                         if ((!i->ref.gun->Dissolved()) || i->ref.gun->Ready()) {
-                            unit->energy.Deplete(true, i->type->energy_rate * simulation_atom_var);
+                            unit->energy.Deplete(true, i->type->energy_rate * static_cast<double>(simulation_atom_var));
                         }
                     }
                 } else if (i->type->isMissile()) {    // FIXME  other than beams, only missiles are processed here?
